Reworked  to support also TLS 1.1 following the implementation included in Facebook React library (PR 9480)
